requirements.txt
ccxt
pandas
pandas_ta
numpy
requests
Flask
gunicorn
python-dotenv
